+++ locals.tf
@@ -1,7 +1,16 @@
 locals {
-  path = "${path.module}/queries"
+  path                    = "${path.module}/queries"
+  service_uri             = replace(var.event_pipeline_config.service_uri, "{{secret}}", var.secret)
+  service_uri_integration = replace(var.event_pipeline_config.service_uri_integration, "{{secret}}", var.secret_integration)
 
   rules = merge({
+    # TODO: Verify with monitoring colleagues
+    "alr-prd-Function-winux-law-logsea-crit-01": {
+      description = "Alert when function app failed"
+      query_path  = "${local.path}/function_app.kusto"
+      time_window = "P2D"
+      frequency   = "PT5M"
+    }
     "alr-prd-Heartbeat-ux-law-metric-crit-01" : {
       description = "Alert when Heartbeat of unix machines Stopped"
       query_path  = "${local.path}/unix_heartbeat.kusto"
@@ -70,4 +79,19 @@ locals {
 
   customer_code = upper(split("-", regex("fctkey-[^-]+", var.event_pipeline_config.service_uri_integration))[1])
 
+  # Add <name> = <directory> entries for each function you want to add to the monitoring function app.
+  # Note: the <name> must correspond to the variable function_config.stage_<name>
+  # Note: the directory is the name of the subdirectory in /functions where the function code is located
+  all_functions = [
+    "mssql"
+  ]
+
+  sql_key_vault_name = format("kv-%s-sqlmonitor-01", local.customer_code)
+
+  # Exclude source code of all functions that are not specified or specifically set to "off"
+  excluded_functions = [for func in local.all_functions : func if lookup(var.functions_config.stages, func, "off") == "off"]
+  enabled_functions = [for func in local.all_functions : func if lookup(var.functions_config.stages, func, "off") != "off"]
+
+  # Checks that the event pipeline and at least one function is enabled
+  enable_function_app = var.event_pipeline_config.enabled && length(local.excluded_functions) < length(local.all_functions)
 }